audio-ingestion-grpc-server/requirements.txt
respeaker
webrtcvad
pocketsphinx
grpcio
grpcio-tools
wave
protobuf
numpy
scipy
